requirements.txt
streamlit
ultralytics
torch
torchvision
pillow
numpy
opencv-python